# Add a minimal global worker pool