--- a/js/global_worker_pool.js
+++ b/js/global_worker_pool.js
@@ -1,0 +1,16 @@
+'use strict';
+var WorkerPool = require('./util/worker_pool');
+
+var globalWorkerPool;
+
+/**
+ * Creates (if necessary) and returns the single, global WorkerPool instance
+ * to be shared across each Map
+ * @private
+ */
+module.exports = function getGlobalWorkerPool () {
+    if (!globalWorkerPool) {
+        globalWorkerPool = new WorkerPool();
+    }
+    return globalWorkerPool;
+};