--- a/js/mapbox-gl.js
+++ b/js/mapbox-gl.js
@@ -18,8 +18,6 @@ mapboxgl.Scale = require('./ui/control/scale');
 mapboxgl.Popup = require('./ui/popup');
 mapboxgl.Marker = require('./ui/marker');
 
-mapboxgl.WorkerPool = require('./util/worker_pool');
-
 mapboxgl.Style = require('./style/style');
 
 mapboxgl.LngLat = require('./geo/lng_lat');