--- a/js/style/style.js
+++ b/js/style/style.js
@@ -18,12 +18,15 @@ var QueryFeatures = require('../source/query_features');
 var SourceCache = require('../source/source_cache');
 var styleSpec = require('./style_spec');
 var StyleFunction = require('./style_function');
+var WorkerPool = require('../util/worker_pool');
+
+var workerPool = new WorkerPool();
 
 module.exports = Style;
 
 function Style(stylesheet, animationLoop, workerCount) {
     this.animationLoop = animationLoop || new AnimationLoop();
-    this.dispatcher = new Dispatcher(workerCount || 1, this);
+    this.dispatcher = new Dispatcher(workerPool, workerCount || 1, this);
     this.spriteAtlas = new SpriteAtlas(1024, 1024);
     this.lineAtlas = new LineAtlas(256, 512);