--- a/js/style/style.js
+++ b/js/style/style.js
@@ -732,7 +732,7 @@ Style.prototype = util.inherit(Evented, {
 
     // Callbacks from web workers
 
-    'get sprite json': function(_, params, callback) {
+    'get sprite json': function(mapId, params, callback) {
         var sprite = this.sprite;
         if (sprite.loaded()) {
             callback(null, { sprite: sprite.data, retina: sprite.retina });
@@ -743,7 +743,7 @@ Style.prototype = util.inherit(Evented, {
         }
     },
 
-    'get icons': function(_, params, callback) {
+    'get icons': function(mapId, params, callback) {
         var sprite = this.sprite;
         var spriteAtlas = this.spriteAtlas;
         if (sprite.loaded()) {
@@ -757,7 +757,7 @@ Style.prototype = util.inherit(Evented, {
         }
     },
 
-    'get glyphs': function(_, params, callback) {
+    'get glyphs': function(mapId, params, callback) {
         var stacks = params.stacks,
             remaining = Object.keys(stacks).length,
             allGlyphs = {};