--- a/js/util/actor.js
+++ b/js/util/actor.js
@@ -10,11 +10,13 @@ module.exports = Actor;
  *
  * @param {WebWorker} target
  * @param {WebWorker} parent
+ * @param {string|number} mapId A unique identifier for the Map instance using this Actor.
  * @private
  */
-function Actor(target, parent) {
+function Actor(target, parent, mapId) {
     this.target = target;
     this.parent = parent;
+    this.mapId = mapId;
     this.callbacks = {};
     this.callbackID = 0;
     this.receive = this.receive.bind(this);
@@ -32,17 +34,19 @@ Actor.prototype.receive = function(message) {
         if (callback) callback(data.error || null, data.data);
     } else if (typeof data.id !== 'undefined' && this.parent[data.type]) {
         // data.type == 'load tile', 'remove tile', etc.
-        this.parent[data.type](data.data, done.bind(this));
-    } else if (typeof data.id !== 'undefined' && this.parent.workerSources) {
+        this.parent[data.type](data.mapId, data.data, done.bind(this));
+    } else if (typeof data.id !== 'undefined' && this.parent.getWorkerSource) {
         // data.type == sourcetype.method
         var keys = data.type.split('.');
-        this.parent.workerSources[keys[0]][keys[1]](data.data, done.bind(this));
+        var workerSource = this.parent.getWorkerSource(data.mapId, keys[0]);
+        workerSource[keys[1]](data.data, done.bind(this));
     } else {
         this.parent[data.type](data.data);
     }
 
     function done(err, data, buffers) {
         this.postMessage({
+            mapId: this.mapId,
             type: '<response>',
             id: String(id),
             error: err ? String(err) : null,
@@ -54,7 +58,7 @@ Actor.prototype.receive = function(message) {
 Actor.prototype.send = function(type, data, callback, buffers) {
     var id = null;
     if (callback) this.callbacks[id = this.callbackID++] = callback;
-    this.postMessage({ type: type, id: String(id), data: data }, buffers);
+    this.postMessage({ mapId: this.mapId, type: type, id: String(id), data: data }, buffers);
 };
 
 /**