Closes #899 

@lucaswoj this is a much more focused / less ambitious change than #2917.  It adds a global worker pool, making only the minimal changes necessary to allow a Worker to serve multiple map instances correctly. 

It does _not_ attempt to deduplicate the work that is done for different map instances.  Nonetheless, there are benefits to sharing workers.  Immediate benefit: the overhead to create a worker is significant, because it essentially involves copying all of the worker code over from the main thread.  Future benefit: using shared workers paves the way for more ambitious performance inter-map-instance improvements like #2951 

## Specific Notes
 - The majority of the +505 lines come from new tests, benchmark, and debug page.
 - Essentially, the simple strategy here is to change the worker's state-keeping objects `Worker#layers`, and `Worker#layerFamilies` into objects keyed off of `mapId`.
 - The WorkerSource abstraction helps here, because it allowed for implementing per-map state management for things like geojson-vt indexes and cached WorkerTiles, but with pretty minimal code changes.  Before this, we instantiated a single `WorkerSource` instance for each source type.  Now,  when a WorkerSource is registered, we just hold on to the class, and instantiate a separate one for each map (or rather, for each `mapId`).  Doing it this way means that none of the code in GeoJSONWorkerSource or VectorTileWorkerSource had to be touched.
 - The `workerCount` Map option is removed in favor of `mapboxgl.WorkerPool.WORKER_COUNT`

## Checklist

 - [x] briefly describe the changes in this PR
 - [x] Add unit test to assert that `Dispatcher` correctly consumes `WorkerPool`
 - [x] Add unit tests for `WorkerPool`
 - [x] Add unit test for two `Actor`s using the same underlying `Worker`.
 - [x] **(n/a)** [document any changes or additions to public APIs](https://github.com/mapbox/mapbox-gl-js/blob/master/docs/README.md)
 - [x] [post scores for all benchmarks on your branch and `master`](https://github.com/mapbox/mapbox-gl-js/blob/master/bench/README.md#running-benchmarks)
 - [x] [do a quick sanity check on the debug page](https://github.com/mapbox/mapbox-gl-js/blob/master/CONTRIBUTING.md#serving-the-debug-page)
 - [x] get a PR review :+1: / :-1: