@anandthakker 2016-08-06
Note that the first commit here is equivalent to #2948