@anandthakker 2016-08-09
Rebasing, previous at https://github.com/mapbox/mapbox-gl-js/pull/2952/commits/d3d0787c835af4b10201bde0d0f7a020011b2da3