@anandthakker 2016-08-09
FWIW: I'm using this branch in a project where we have multiple maps on a page, and even just having a shared pool--without actually deduping any of the work--significantly reduces the initial load time, because initializing each Worker is a pretty significant overhead.