@anandthakker 2016-08-09
Next steps:
 - [ ] 👍 / 👎 from @lucaswoj 
 - [ ] Add unit test to assert that `Dispatcher` correctly consumes `WorkerPool`
 - [ ] Add unit tests for `WorkerPool`