@anandthakker 2016-08-11
Updated top of ticket with rationale and the checklist from the new PR template