@anandthakker 2016-08-12
Wouldn't expect these changes to affect the benchmarks much, but why not:

## This branch

**buffer:** 886 ms
**fps:** 60 fps
**frame-duration:** 6.9 ms, 0% > 16ms
**query-point:** 0.64 ms
**query-box:** 47.80 ms
**geojson-setdata-small:** 13 ms
**geojson-setdata-large:** 115 ms

## Master

Broke w/ console errors at the end of query-point, but here are the first
three:

**buffer:** 868 ms
**fps:** 58 fps
**frame-duration:** 6.8 ms, 0% > 16ms