@anandthakker 2016-08-12
Added a benchmark to measure the time from `new Map` to the `load` event, with multiple maps on a page.

## master

**load-multiple-maps:** 7,024 ms, loaded 6 maps.

## This branch

**load-multiple-maps:** 4,778 ms, loaded 6 maps.