@anandthakker 2016-08-18
In light of a global worker pool, I don't know that either of the following seem necessary/useful: (a) having different map instances use a different number of workers, or (b) being able to resize the worker pool.

Does it make sense to remove the `workerCount` option from `Map` (and `Style` and `Dispatcher`)?  Instead there could be an "advanced" API like `mapboxgl.workerPool.setWorkerCount()`, which would need to be called before any map instances were constructed.