@mourner 2016-08-18
Makes sense. I would go for something like `mapboxgl.WORKER_COUNT` — it would only be changed for development/debugging purposes anyway.