@anandthakker 2016-08-18
> I would go for something like mapboxgl.WORKER_COUNT — it would only be changed for development/debugging purposes anyway.

I think it could be set by a user who was tying to balance a GL JS map versus other cpu-intensive stuff in a web app.  (For instance: in a project we're working on, I was thinking of intentionally limiting the GL JS worker count to 3/4 of the available workers, so that 1/4 could be dedicated to keeping some other expensive data-munging off the main thread to keep the UI snappier)