@anandthakker 2016-08-18
^ but, with that said, I don't actually have any objections to making it a constant like you proposed, @mourner