@anandthakker 2016-08-18
Actually, just a note that making this a constant on the `mapboxgl` object causes some weirdness, because worker_pool then has to depend on the parent mapboxgl module.  Means having to be careful about a dependency cycle, and I think might also make unit tests a bit more awkward.  How about `mapboxgl.workerPool.WORKER_COUNT`?   @mourner