@anandthakker 2016-08-19
Went with `mapboxgl.WorkerPool.WORKER_COUNT` - @mourner let me know what you think.

(note: also rebased to master)