@anandthakker 2016-08-19
Updated the top of the ticket.  @lucaswoj while I'm very much in favor of https://github.com/mapbox/mapbox-gl-js/issues/3034, I would propose that the changes in this PR  provide some immediate benefits and incur little to no cost.  What do you think?