@lucaswoj 2016-08-19
> Went with `mapboxgl.WorkerPool.WORKER_COUNT` - @mourner let me know what you think.

Let's go with `mapboxgl.workerCount` to match `mapboxgl.accessToken`. We can avoid cyclic dependencies by lazily calling `require('../../mapbox-gl').workerCount` when the value is needed.