@anandthakker 2016-08-21
@lucaswoj addressed all your notes except the one about scoping the worker pool to the style.js file, where there may be a reason to keep it as-is: https://github.com/mapbox/mapbox-gl-js/pull/2952#discussion_r75588490