@anandthakker 2016-08-22
Heads up -- rebased to (hopefully) make coveralls happy again