@anandthakker 2016-08-22
Ugh, I think the reason for the coveralls failure is that, for some reason, `mapbox-gl.js` (and some of the un-tested modules it requires, like control/geolocate.js, control/navigation.js, etc.) are _not_ included in the list of source files in the [`master` build](https://coveralls.io/github/mapbox/mapbox-gl-js), but _are_ included in the list for [this branch](https://coveralls.io/builds/7550387).

This is strange, because `mapbox-gl.js` does have a test in master, so it should be getting included.  My guess is that the difference between the two branches has something to do with using `proxyquire` in this branch to mock `mapboxgl.workerCount` -- but, again, mapbox-gl *should* have already been in the list of files being checked for coverage.