@anandthakker 2016-08-22
> This is strange, because mapbox-gl.js does have a test in master, so it should be getting included. My guess is that the difference between the two branches has something to do with using proxyquire in this branch to mock mapboxgl.workerCount -- but, again, mapbox-gl should have already been in the list of files being checked for coverage.

Ah - it's a simple problem.  The `npm test` script runs tap on `test/js/*/*.js`, but the mapbox-gl test is at `test/js/mapbox-gl.js`.   See https://github.com/mapbox/mapbox-gl-js/pull/3043