@lucaswoj 2016-08-22
This is ready to merge on 🍏 . Just rebased to include #3043.