@anandthakker 2016-08-23
Rebasing -- previous at https://github.com/mapbox/mapbox-gl-js/pull/2952/commits/84ce97bdd44c6bc72508f482f28668f05a30c0b8