@anandthakker 2016-08-25
Ack - just realized that I dropped the ball on [this comment](https://github.com/mapbox/mapbox-gl-js/pull/2952#discussion_r75548284) -- the diff changed while I was addressing something else, and so I forgot about it.

@lucaswoj if I don't get a chance to open a PR patching that today, I'll file an issue so it doesn't get lost